side_by_side_diff: move util to test
