+++ onlinejudge/_implementation/utils.py
@@ -5,11 +5,9 @@ import distutils.version
 import functools
 import http.client
 import http.cookiejar
-import itertools
 import json
 import os
 import posixpath
-import re
 import shlex
 import shutil
 import subprocess
@@ -21,10 +19,8 @@ from typing import *
 from typing.io import *
 
 import bs4
-import diff_match_patch
 
 import onlinejudge.__about__ as version
-import onlinejudge._implementation.logging as log
 import onlinejudge.utils
 from onlinejudge.type import *
 from onlinejudge.utils import *  # re-export
@@ -349,86 +345,3 @@ class DummySubmission(Submission):
     @classmethod
     def from_url(cls, s: str) -> Optional[Submission]:
         return None
-
-
-def yield_open_entry(open_entry: Tuple[List[str], List[str], List[int], List[int]]) -> Generator[Tuple[bool, str, str, int, int], None, None]:
-    """ Yield all open changes. """
-    ls, rs, lnums, rnums = open_entry
-    # Get unchanged parts onto the right line
-    if ls[0] == rs[0]:
-        yield (False, ls[0], rs[0], lnums[0], rnums[0])
-        for l, r, lnum, rnum in itertools.zip_longest(ls[1:], rs[1:], lnums[1:], rnums[1:]):
-            yield (True, l or '', r or '', lnum or 0, rnum or 0)
-    elif ls[-1] == rs[-1]:
-        for l, r, lnum, rnum in itertools.zip_longest(ls[:-1], rs[:-1], lnums[:-1], rnums[:-1]):
-            yield (l != r, l or '', r or '', lnum or 0, rnum or 0)
-        yield (False, ls[-1], rs[-1], lnums[-1], rnums[-1])
-    else:
-        for l, r, lnum, rnum in itertools.zip_longest(ls, rs, lnums, rnums):
-            yield (True, l or '', r or '', lnum or 0, rnum or 0)
-
-
-def side_by_side_diff(old_text: str, new_text: str) -> Generator[Tuple[bool, str, str, int, int], None, None]:
-    """
-    Calculates a side-by-side line-based difference view.
-    """
-    line_split = re.compile(r'(?:\r?\n)')
-    dmp = diff_match_patch.diff_match_patch()
-
-    diff = dmp.diff_main(old_text, new_text)
-    dmp.diff_cleanupSemantic(diff)
-
-    open_entry = ([''], [''], [0], [0])
-    for change_type, entry in diff:
-        assert change_type in [-1, 0, 1]
-
-        entry = (entry.replace('&', '&amp;').replace('<', '&lt;').replace('>', '&gt;'))
-        lines = line_split.split(entry)
-
-        # Merge with previous entry if still open
-        ls, rs, lnums, rnums = open_entry
-
-        line = lines[0]
-        if line:
-            if change_type == 0:
-                ls[-1] += line
-                rs[-1] += line
-                lnums[-1] += len(line)
-                rnums[-1] += len(line)
-            elif change_type == 1:
-                rs[-1] = rs[-1] or ''
-                rs[-1] += log.green_diff(line) if line else ''
-                rnums[-1] += len(line)
-            elif change_type == -1:
-                ls[-1] = ls[-1] or ''
-                ls[-1] += log.red_diff(line) if line else ''
-                lnums[-1] += len(line)
-
-        lines = lines[1:]
-
-        if lines:
-            if change_type == 0:
-                # Push out open entry
-                for entry in yield_open_entry(open_entry):
-                    yield entry
-
-                # Directly push out lines until last
-                for line in lines[:-1]:
-                    yield (False, line, line, len(line), len(line))
-
-                # Keep last line open
-                open_entry = ([lines[-1]], [lines[-1]], [len(lines[-1])], [len(lines[-1])])
-            elif change_type == 1:
-                ls, rs, lnums, rnums = open_entry
-                for line in lines:
-                    rs.append(log.green_diff(line) if line else '')
-                    rnums.append(len(line))
-            elif change_type == -1:
-                ls, rs, lnums, rnums = open_entry
-                for line in lines:
-                    ls.append(log.red_diff(line) if line else '')
-                    lnums.append(len(line))
-
-    # Push out open entry
-    for entry in yield_open_entry(open_entry):
-        yield entry
